Code refactor and review
